修正image的bug

+++ RentPlace/src/main/java/fcu/iecs/demo/repository/ImageRepository.java
@@ -3,10 +3,5 @@ package fcu.iecs.demo.repository;
 import fcu.iecs.demo.model.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
-
-public interface ImageRepository extends JpaRepository<Image, Long> {
-
+public interface ImageRepository extends JpaRepository<Image, Integer> {
 }
-
-
